+++ lib/domain/models/time_series.dart
@@ -3,7 +3,7 @@ import 'package:vsd/domain/models/statistic.dart';
 class DataPoint {
   DataPoint({required this.timestamp, required this.value});
   DateTime timestamp;
-  double value;
+  int value;
 }
 
 class TimeSeries {
@@ -17,7 +17,7 @@ class TimeSeries {
 
   // Helper methods
   bool get hasData => points.isNotEmpty && points.any((p) => p.value != 0);
-  double? get min => points.isEmpty ? null : points.map((p) => p.value).reduce((a, b) => a < b ? a : b);
-  double? get max => points.isEmpty ? null : points.map((p) => p.value).reduce((a, b) => a > b ? a : b);
+  int? get min => points.isEmpty ? null : points.map((p) => p.value).reduce((a, b) => a < b ? a : b);
+  int? get max => points.isEmpty ? null : points.map((p) => p.value).reduce((a, b) => a > b ? a : b);
   double? get average => points.isEmpty ? null : points.map((p) => p.value).reduce((a, b) => a + b) / points.length;
 }
